Rewrite makefile so it only rebuilds what is necessary
The current makefile doesn't correctly define the dependencies of the
various binaries.  This means that if you run 'make' it will rebuild
everything even if no source files have changed.  Restructure the makefile
to define the dependencies so that we only build what we need to.

The key changes here are:
 * use an 'order-only dependency' for ./bin/ so we don't rebuild
   everything when the directory's timestamp changes
 * use the actual binary filenames rather than phony target names for
   the rules that build the binaries
 * mark that the binaries depend on the source files that go into them
 * use Make's automatic variables (like $^) to avoid repeating the
   list of source files in the rule itself